+++ Tasvee/src/main/java/fr/pantheonsorbonne/ufr27/miage/dao/OfferFormDAO.java
@@ -1,0 +1,7 @@
+package fr.pantheonsorbonne.ufr27.miage.dao;
+
+import fr.pantheonsorbonne.ufr27.miage.dto.OfferForm;
+
+public interface OfferFormDAO {
+    void saveOfferForm(OfferForm offerForm) ;
+}
