+++ Tasvee/src/main/java/fr/pantheonsorbonne/ufr27/miage/service/OfferFormService.java
@@ -3,7 +3,5 @@ package fr.pantheonsorbonne.ufr27.miage.service;
 import fr.pantheonsorbonne.ufr27.miage.dto.OfferForm;
 
 public interface OfferFormService {
-    void saveOfferForm(OfferForm offerForm) ;
-
     boolean isOfferAccepted(OfferForm offerForm) ;
 }
